# docs/api/gateway.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PHS — API pública (gateway)",
    "version": "0.1.0",
    "description": "Contrato que usa la aplicación web. Toda ruta entra por el gateway, que valida la sesión y reenvía a los servicios."
  },
  "paths": {
...
    "/api/v1/session": {
      "post": {
        "summary": "Iniciar sesión",
        "tags": [
          "Sesión"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 320
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada. El token viaja solo en la cookie `phs_session` (HttpOnly, SameSite=Strict).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "mustChangePassword": {
                          "type": "boolean",
                          "description": "La contraseña es temporal y debe cambiarse antes de operar."
                        },
                        "isAdmin": {
                          "type": "boolean"
                        },
                        "memberships": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "practiceId": {
                                "type": "string",
                                "format": "uuid",
                                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                              },
                              "practiceName": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string",
                                "enum": [
                                  "pm",
                                  "lead",
                                  "director"
                                ],
                                "description": "Rol de un usuario en una práctica."
                              }
                            },
                            "required": [
                              "practiceId",
                              "practiceName",
                              "role"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "id",
                        "displayName",
                        "email",
                        "mustChangePassword",
                        "isAdmin",
                        "memberships"
                      ],
                      "additionalProperties": false
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "user",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (`invalid_request`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas o cuenta bloqueada (`invalid_credentials`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido (`origin_not_allowed`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Sesión actual",
        "tags": [
          "Sesión"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario, roles y vencimiento.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "mustChangePassword": {
                          "type": "boolean",
                          "description": "La contraseña es temporal y debe cambiarse antes de operar."
                        },
                        "isAdmin": {
                          "type": "boolean"
                        },
                        "memberships": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "practiceId": {
                                "type": "string",
                                "format": "uuid",
                                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                              },
                              "practiceName": {
                                "type": "string"
                              },
                              "role": {
                                "type": "string",
                                "enum": [
                                  "pm",
                                  "lead",
                                  "director"
                                ],
                                "description": "Rol de un usuario en una práctica."
                              }
                            },
                            "required": [
                              "practiceId",
                              "practiceName",
                              "role"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "id",
                        "displayName",
                        "email",
                        "mustChangePassword",
                        "isAdmin",
                        "memberships"
                      ],
                      "additionalProperties": false
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    }
                  },
                  "required": [
                    "user",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión válida (`authentication_required`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cerrar sesión",
        "tags": [
          "Sesión"
        ],
        "security": [],
        "responses": {
          "204": {
            "description": "Sesión revocada y cookie eliminada. Responde igual si no había sesión."
          }
        }
      }
    },
    "/api/v1/session/password": {
      "post": {
        "summary": "Cambiar la contraseña propia",
        "tags": [
          "Sesión"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024,
                    "description": "Entre 12 y 128 caracteres y distinta de la actual."
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Contraseña cambiada."
          },
          "400": {
            "description": "Datos inválidos (`invalid_request`) o contraseña no aceptada (`weak_password`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión (`authentication_required`) o contraseña actual incorrecta (`invalid_current_password`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "El servicio de identidad no respondió (`identity_unavailable`).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Código estable del error, pensado para que el cliente decida el mensaje."
                    }
                  },
                  "required": [
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "phs_session",
        "description": "Cookie de sesión emitida por POST /api/v1/session."
      },
      "gatewaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-phs-internal-auth",
        "description": "Identidad firmada por el gateway (HMAC-SHA256, válida 60 s). No se acepta desde el navegador."
      }
    }
  }
}